refactor(hfb): move hfb into new subpackage
