+++ ch_pipeline/hfb/__init__.py
@@ -1,0 +1,12 @@
+"""
+CHIME HFB Pipeline
+
+.. currentmodule:: ch_pipeline.hfb
+
+.. autosummary::
+    :toctree:
+
+    containers
+    io
+
+"""
