+++ ch_pipeline/hfb/io.py
@@ -1,0 +1,99 @@
+"""HFB-related tasks
+"""
+from typing import Union
+
+from caput import tod
+from ch_util import andata
+
+from .containers import HFBData
+
+
+class HFBReader(tod.Reader):
+    """A reader for HFB type data."""
+
+    data_class = HFBData
+
+    _freq_sel = None
+
+    @property
+    def freq_sel(self) -> Union[int, list, slice]:
+        """Get the current frequency selection.
+
+        Returns
+        -------
+        freq_sel
+            A frequency selection.
+        """
+
+        return self._freq_sel
+
+    @freq_sel.setter
+    def freq_sel(self, value: Union[int, list, slice]):
+        """Set a frequency selection.
+
+        Parameters
+        ----------
+        value
+            Any type accepted by h5py is valid.
+        """
+        self._freq_sel = andata._ensure_1D_selection(value)
+
+    _beam_sel = None
+
+    @property
+    def beam_sel(self):
+        """Get the current beam selection.
+
+        Returns
+        -------
+        beam_sel
+            The current beam selection.
+        """
+
+        return self._beam_sel
+
+    @beam_sel.setter
+    def beam_sel(self, value):
+        """Set a beam selection.
+
+        Parameters
+        ----------
+        value
+            Any type accepted by h5py is valid.
+        """
+        self._beam_sel = andata._ensure_1D_selection(value)
+
+    def read(self, out_group=None):
+        """Read the selected data.
+
+        Parameters
+        ----------
+        out_group : `h5py.Group`, hdf5 filename or `memh5.Group`
+            Underlying hdf5 like container that will store the data for the
+            BaseData instance.
+
+        Returns
+        -------
+        data : :class:`TOData`
+            Data read from :attr:`~Reader.files` based on the selections made
+            by user.
+
+        """
+        kwargs = {}
+
+        if self._freq_sel is not None:
+            kwargs["freq_sel"] = self._freq_sel
+
+        if self._beam_sel is not None:
+            kwargs["beam_sel"] = self._beam_sel
+
+        kwargs["ondisk"] = False
+
+        return self.data_class.from_mult_files(
+            self.files,
+            data_group=out_group,
+            start=self.time_sel[0],
+            stop=self.time_sel[1],
+            datasets=self.dataset_sel,
+            **kwargs,
+        )
